requirements_stable.txt
# For GUI support, make sure the system package is installed:
# Ubuntu/Debian: sudo apt install python3-tk

# Core packages
numpy>=1.26,<2.4
pandas>=2.0.3,<2.2
matplotlib>=3.7.5,<3.9
pyparsing>=3.0.9,<3.2
pillow>=10.4.0,<11
torch==2.4.1
torchvision==0.19.1

# Networking / HTTP
aiohttp>=3.10.10,<3.11
async-timeout>=4.0,<5.0

# Data packages
datasets>=3.1.0,<3.2
fsspec>=2023.1.0,<2025.0
pyarrow>=17.0.0,<18.0
ruamel.yaml>=0.17.32,<1.0
scikit-learn>=1.5.2,<1.6

# Federated learning frameworks
flwr>=1.11.1,<1.12
cryptography>=42.0.4,<43.0.0

# Utility packages
dill>=0.3.0,<0.3.9
multiprocess>=0.70.0,<0.70.17
zipp>=3.20.0,<3.21.0
yarl>=1.15.2,<1.18.3
tomli_w>=1.0.0,<1.1.0
rpds-py>=0.20.1,<0.22.3

omegaconf>=2.3.0,<2.4

# Optional but useful
networkx==3.4.2
ray>=2.10.0,<2.30

